+++ mmpose/datasets/datasets/top_down/topdown_onehand10k_dataset.py
@@ -108,8 +108,8 @@ class TopDownOneHand10KDataset(TopDownBaseDataset):
             if max(anno['keypoints']) == 0:
                 continue
 
-            joints_3d = np.zeros((num_joints, 3), dtype=np.float)
-            joints_3d_visible = np.zeros((num_joints, 3), dtype=np.float)
+            joints_3d = np.zeros((num_joints, 3), dtype=np.float32)
+            joints_3d_visible = np.zeros((num_joints, 3), dtype=np.float32)
 
             for ipt in range(num_joints):
                 joints_3d[ipt, 0] = anno['keypoints'][ipt * 3 + 0]
